// src/app/models/PagingParams.ts
export interface PagingParams {
    PageNumber?: number;
    PageSize?: number;
    Keyword?: string;
    SortValue?: string;
    SortKey?: string;
    fromDate?: string;
    toDate?: string;
    KeywordCol?: string;
    ColName?: string;
    userId?: string;
    type?: string;
    Filter?: any[];
    TrangThaiDoanhNghiep?: any;
    LoaiTieuChi?: number;
    LoaiCanhBao?: number;
}